requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==1.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.0.0
redis==5.0.1
elasticsearch==8.9.0
bcrypt==4.0.1
celery==5.3.4